# implement: vectorized sigmoid, sigmoid_prime