@maxpumperla 2019-02-11
@jeffhgs thanks for your addition. Actually, we're completely aware of this. Two things:

a) Chapter 5 is "optimized" for readability, aimed at beginners. I hope you see how your performance bump is a little less readable for someone getting started.
b) The master branch should stay in sync with the book. Exceptions: clarifications and bug fixes / errata.

Having said that, I don't want your work to go to waste. I'm thinking about having an `improvements` branch people can open PRs against. Would that work for you? I'll put this on the readme as well. :+1: