@jeffhgs 2019-02-11
Reading:

> Note for contributors: To ensure the book stays in sync, consider requesting changes and submitting pull requests against the improvements branch, instead of master (which we keep reserved for bug fixes etc.).

Sure, a branching policy to preserve book sync totally seems appropriate.